+++ framework/decode/openxr_replay_consumer_base.cpp
@@ -1046,7 +1046,7 @@ void OpenXrReplayConsumerBase::Process_xrEnumerateDisplayRefreshRatesFB(
             // Allocate the blend mode array and get all the values
             std::vector<float> display_refresh_rates(replay_count);
             replay_result = pfn_enum_dis_refresh_rates_fb(in_session,
-                                                          display_refresh_rates.size(),
+                                                          static_cast<uint32_t>(display_refresh_rates.size()),
                                                           out_displayRefreshRateCountOutput,
                                                           display_refresh_rates.data());
