+++ framework/decode/vulkan_replay_dump_resources_draw_calls.cpp
@@ -111,7 +111,7 @@ void DrawCallsDumpingContext::Release()
         if (!command_buffers_.empty())
         {
             device_table_->FreeCommandBuffers(
-                device, pool_info->handle, command_buffers_.size(), command_buffers_.data());
+                device, pool_info->handle, static_cast<uint32_t>(command_buffers_.size()), command_buffers_.data());
         }
         command_buffers_.clear();
 
@@ -327,7 +327,7 @@ VkResult DrawCallsDumpingContext::CopyDrawIndirectParameters(DrawCallParams& dc_
             device_table_->CmdCopyBuffer(cmd_buf,
                                          ic_params.params_buffer_info->handle,
                                          ic_params.new_params_buffer,
-                                         regions.size(),
+                                         static_cast<uint32_t>(regions.size()),
                                          regions.data());
 
             VkBufferMemoryBarrier buf_barrier;
@@ -470,7 +470,7 @@ VkResult DrawCallsDumpingContext::CopyDrawIndirectParameters(DrawCallParams& dc_
             device_table_->CmdCopyBuffer(cmd_buf,
                                          i_params.params_buffer_info->handle,
                                          i_params.new_params_buffer,
-                                         regions.size(),
+                                         static_cast<uint32_t>(regions.size()),
                                          regions.data());
 
             VkBufferMemoryBarrier buf_barrier;
@@ -969,7 +969,7 @@ VkResult DrawCallsDumpingContext::RevertRenderTargetImageLayouts(VkQueue queue, 
         return VK_SUCCESS;
     }
 
-    const auto entry = rendering_attachment_layouts_.find(rp);
+    const auto entry = rendering_attachment_layouts_.find(static_cast<uint32_t>(rp));
     assert(entry != rendering_attachment_layouts_.end());
 
     if (!entry->second.is_dynamic)
@@ -1041,7 +1041,7 @@ VkResult DrawCallsDumpingContext::RevertRenderTargetImageLayouts(VkQueue queue, 
                                           nullptr,
                                           0,
                                           nullptr,
-                                          img_barriers.size(),
+                                          static_cast<uint32_t>(img_barriers.size()),
                                           img_barriers.data());
 
         res = device_table_->EndCommandBuffer(aux_command_buffer_);
@@ -1986,10 +1986,10 @@ VkResult DrawCallsDumpingContext::DumpVertexIndexBuffers(uint64_t qs_index, uint
         {
             const VkIndexType index_type = dc_params.referenced_index_buffer.index_type;
             const uint32_t    index_size = VkIndexTypeToBytes(index_type);
-            const uint32_t    offset     = dc_params.referenced_index_buffer.offset;
+            const uint64_t    offset     = dc_params.referenced_index_buffer.offset;
 
             // Check if the exact size has been provided by vkCmdBindIndexBuffer2
-            uint32_t total_size = (dc_params.referenced_index_buffer.size != 0)
+            uint64_t total_size = (dc_params.referenced_index_buffer.size != 0)
                                       ? (dc_params.referenced_index_buffer.size)
                                       : (abs_index_count * index_size);
 
@@ -2196,8 +2196,8 @@ VkResult DrawCallsDumpingContext::DumpVertexIndexBuffers(uint64_t qs_index, uint
 
                 const uint32_t count =
                     binding_desc.inputRate == VK_VERTEX_INPUT_RATE_VERTEX ? vertex_count : instance_count;
-                uint32_t total_size = 0;
-                uint32_t binding_stride;
+                uint64_t total_size = 0;
+                uint64_t binding_stride;
 
                 if (vb_entry.size)
                 {
@@ -2227,8 +2227,10 @@ VkResult DrawCallsDumpingContext::DumpVertexIndexBuffers(uint64_t qs_index, uint
                                 continue;
                             }
 
-                            total_size = std::max(
-                                total_size, vkuFormatElementSize(input_attrib_desc.format) + input_attrib_desc.offset);
+                            total_size =
+                                std::max(total_size,
+                                         static_cast<uint64_t>(vkuFormatElementSize(input_attrib_desc.format)) +
+                                             input_attrib_desc.offset);
                         }
 
                         if (!total_size)
@@ -2239,7 +2241,7 @@ VkResult DrawCallsDumpingContext::DumpVertexIndexBuffers(uint64_t qs_index, uint
                 }
 
                 // Calculate offset including vertexOffset
-                uint32_t offset = vb_entry.offset;
+                uint64_t offset = vb_entry.offset;
                 offset +=
                     (binding_desc.inputRate == VK_VERTEX_INPUT_RATE_VERTEX ? min_max_vertex_indices.min + first_vertex
                                                                            : first_instance) *
@@ -2385,7 +2387,7 @@ void DrawCallsDumpingContext::BindDescriptorSets(
     }
 
     uint32_t dynamic_offset_index = 0;
-    for (size_t i = 0; i < descriptor_sets_infos.size(); ++i)
+    for (uint32_t i = 0; i < static_cast<uint32_t>(descriptor_sets_infos.size()); ++i)
     {
         const uint32_t set_index = first_set + i;
 
@@ -2536,14 +2538,14 @@ VkResult DrawCallsDumpingContext::CloneRenderPass(const VkRenderPassCreateInfo* 
         VkRenderPassCreateInfo ci;
         ci.sType           = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO;
         ci.flags           = original_render_pass_ci->flags;
-        ci.attachmentCount = modified_attachments.size();
+        ci.attachmentCount = static_cast<uint32_t>(modified_attachments.size());
         ci.pAttachments    = modified_attachments.empty() ? nullptr : modified_attachments.data();
 
         assert(subpass_descs.size() == sub + 1);
         ci.subpassCount = sub + 1;
         ci.pSubpasses   = subpass_descs.data();
 
-        ci.dependencyCount = modified_dependencies.size();
+        ci.dependencyCount = static_cast<uint32_t>(modified_dependencies.size());
         ci.pDependencies   = modified_dependencies.empty() ? nullptr : modified_dependencies.data();
 
         ci.pNext = original_render_pass_ci->pNext;
@@ -2691,14 +2693,14 @@ VkResult DrawCallsDumpingContext::CloneRenderPass2(const VulkanRenderPassInfo*  
         // it doesn't matter which one we use
         ci.sType           = VK_STRUCTURE_TYPE_RENDER_PASS_CREATE_INFO_2;
         ci.flags           = original_render_pass_ci->flags;
-        ci.attachmentCount = modified_attachments.size();
+        ci.attachmentCount = static_cast<uint32_t>(modified_attachments.size());
         ci.pAttachments    = modified_attachments.empty() ? nullptr : modified_attachments.data();
 
         assert(subpass_descs.size() == sub + 1);
         ci.subpassCount = sub + 1;
         ci.pSubpasses   = subpass_descs.data();
 
-        ci.dependencyCount = modified_dependencies.size();
+        ci.dependencyCount = static_cast<uint32_t>(modified_dependencies.size());
         ci.pDependencies   = modified_dependencies.empty() ? nullptr : modified_dependencies.data();
 
         ci.correlatedViewMaskCount = original_render_pass_ci->correlatedViewMaskCount;
@@ -3564,7 +3566,7 @@ uint32_t DrawCallsDumpingContext::GetDrawCallActiveCommandBuffers(CommandBufferI
     GFXRECON_ASSERT(current_cb_index_ <= command_buffers_.size());
     first = command_buffers_.begin() + static_cast<int>(current_cb_index_);
     last  = command_buffers_.end();
-    return current_cb_index_;
+    return static_cast<uint32_t>(current_cb_index_);
 }
 
 void DrawCallsDumpingContext::BeginRendering(const std::vector<VulkanImageInfo*>& color_attachments,
@@ -3616,7 +3618,7 @@ void DrawCallsDumpingContext::AssignSecondary(uint64_t                 execute_c
 
 uint32_t DrawCallsDumpingContext::RecaclulateCommandBuffers()
 {
-    uint32_t n_command_buffers = command_buffers_.size();
+    uint32_t n_command_buffers = static_cast<uint32_t>(command_buffers_.size());
 
     if (secondaries_.empty())
     {
@@ -3627,7 +3629,7 @@ uint32_t DrawCallsDumpingContext::RecaclulateCommandBuffers()
     {
         for (auto& secondary_context : execute_commands.second)
         {
-            const size_t secondary_n_command_buffers = secondary_context->RecaclulateCommandBuffers();
+            const uint32_t secondary_n_command_buffers = secondary_context->RecaclulateCommandBuffers();
             if (!secondary_n_command_buffers)
             {
                 return n_command_buffers;
@@ -3723,7 +3725,7 @@ void DrawCallsDumpingContext::MergeRenderPasses(const DrawCallsDumpingContext& s
         }
     }
 
-    current_renderpass_ += secondary_context.rendering_attachment_layouts_.size();
+    current_renderpass_ += static_cast<uint32_t>(secondary_context.rendering_attachment_layouts_.size());
 }
 
 void DrawCallsDumpingContext::UpdateSecondaries()
